Improve CLI output, auth storage, and config UX
Documentation:
- Update README auth/config guidance and examples

Changes:
- Add output formatting controls and JSON responses
- Default tokens to config file with opt-in keychain access
- Introduce config command and refine compare/export/token flows

+++ src/api/images.rs
@@ -20,7 +20,7 @@ impl FigmaClient {
         file_key: &str,
         node_ids: &[String],
         format: &str,
-        scale: u8,
+        scale: f32,
     ) -> Result<ImageResponse> {
         // Create cache key based on all parameters
         let params_hash = CacheKey::hash_export_params(node_ids, format, scale);
